Strip down logger
