+++ snorkel/labeling/model/baselines.py
@@ -1,6 +1,29 @@
 import numpy as np
+import torch
+from scipy.sparse import issparse
 
-from .label_model import LabelModel, to_numpy
+from .label_model import LabelModel
+
+
+def to_numpy(Z):
+    """Converts a None, list, np.ndarray, or torch.Tensor to np.ndarray;
+    also handles converting sparse input to dense."""
+    if Z is None:
+        return Z
+    elif issparse(Z):
+        return Z.toarray()
+    elif isinstance(Z, np.ndarray):
+        return Z
+    elif isinstance(Z, list):
+        return np.array(Z)
+    elif isinstance(Z, torch.Tensor):
+        return Z.cpu().numpy()
+    else:
+        msg = (
+            f"Expected None, list, numpy.ndarray or torch.Tensor, "
+            f"got {type(Z)} instead."
+        )
+        raise Exception(msg)
 
 
 class BaselineVoter(LabelModel):
